requirements.txt
streamlit
pandas
qrcode
pillow
openpyxl
python-dotenv